requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
multitasking==0.0.9
numpy==1.18.1
pandas==0.25.3
patsy==0.5.1
psycopg2==2.8.5
pylint==2.5.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
statsmodels==0.10.2
toml==0.10.1
urllib3==1.25.9
Werkzeug==0.16.0
wrapt==1.12.1
yfinance==0.1.54
